requirements.txt
# requirements. 
# Use
# pip3 install -r requirements.txt

python-docx
docx2pdf